Refactor code design to reduce coupling between classes
