+++ modi/_executor_task.py
@@ -2,8 +2,6 @@
 
 """Json Excute module."""
 
-from __future__ import absolute_import
-
 import time
 import json
 import queue
@@ -25,8 +23,13 @@ from modi.module import (
     ultrasonic,
 )
 
+from modi.module.module import Module
 
-class ExcutableTask(object):
+
+class ExecutorTask:
+    """
+    This task execute incoming commands
+    """
 
     # variables shared across all class instances
     module_categories = ["network", "input", "output"]
@@ -36,13 +39,16 @@ class ExcutableTask(object):
         "output": ["display", "motor", "led", "speaker"],
     }
 
-    def __init__(self, serial_write_q, recv_q, module_ids, modules, command):
-        super(ExcutableTask, self).__init__()
+    def __init__(self, serial_write_q, recv_q, module_ids, modules):
+        super(ExecutorTask, self).__init__()
         self._serial_write_q = serial_write_q
         self._recv_q = recv_q
         self._module_ids = module_ids
         self._modules = modules
-        self._command = command
+
+        # TODO: comment
+        time.sleep(2)
+        # self.__init_modules()
 
     def start_thread(self):
         try:
@@ -63,31 +69,27 @@ class ExcutableTask(object):
 
     def __update_health(self, msg):
         module_id = msg["s"]
-        current_time_ms = int(time.time() * 1000)
+        curr_time_ms = int(time.time() * 1000)
         msg_decoded = bytearray(base64.b64decode(msg["b"]))
 
         self._module_ids[module_id] = self._module_ids.get(module_id, dict())
-        self._module_ids[module_id]["timestamp"] = current_time_ms
+        self._module_ids[module_id]["timestamp"] = curr_time_ms
         self._module_ids[module_id]["uuid"] = self._module_ids[module_id].get(
             "uuid", str()
         )
         self._module_ids[module_id]["battery"] = int(msg_decoded[3])
 
         if not self._module_ids[module_id]["uuid"]:
-            msg_to_write = self._command.request_uuid(
-                module_id, is_network_module=False
-            )
+            msg_to_write = self.__request_uuid(module_id, is_network_module=False)
             self._serial_write_q.put(msg_to_write)
-            msg_to_write = self._command.request_uuid(module_id, is_network_module=True)
+            msg_to_write = self.__request_uuid(module_id, is_network_module=True)
             self._serial_write_q.put(msg_to_write)
 
-        for module_id, info in list(self._module_ids.items()):
-            if current_time_ms - info["timestamp"] > 2000:
+        for module_id, module_info in list(self._module_ids.items()):
+            if curr_time_ms - module_info["timestamp"] > 2000:
                 for module in self._modules:
-                    if module.uuid == info["uuid"]:
-                        module.set_connection_state(
-                            state=module.ConnectionState.CONNECTED
-                        )
+                    if module.uuid == module_info["uuid"]:
+                        module.set_connection_state(state=True)
                         print("disconnecting : ", module)
 
     def __update_modules(self, msg):
@@ -126,7 +128,7 @@ class ExcutableTask(object):
         # handling re-connected modules
         for module in self._modules:
             if module.uuid == module_uuid and not module.connected:
-                module.set_connection_state(state=module.ConnectionState.CONNECTED)
+                module.set_connection_state(state=True)
 
         # handling newly-connected modules
         if not next(
@@ -137,10 +139,8 @@ class ExcutableTask(object):
                 module_instance = module_template(
                     module_id, module_uuid, self, self._serial_write_q
                 )
-
                 self.__set_pnp(
-                    module_id=module_instance.id,
-                    module_pnp_state=self._command.ModulePnp.OFF,
+                    module_id=module_instance.id, module_pnp_state=Module.ModulePnp.OFF
                 )
                 self._modules.append(module_instance)
                 self._modules.sort(key=lambda module: module.uuid)
@@ -162,10 +162,12 @@ class ExcutableTask(object):
         return module
 
     def __update_property(self, msg):
+        # TODO: comment
         property_number = msg["d"]
         if property_number == 0 or property_number == 1:
             return
 
+        # TODO: comment
         for module in self._modules:
             if module.id == msg["s"]:
                 decoded = bytearray(base64.b64decode(msg["b"]))
@@ -175,24 +177,82 @@ class ExcutableTask(object):
                 )
 
     def __set_pnp(self, module_id, module_pnp_state):
-        # pnp_state = (
-        #    self._command.ModulePnp.ON if pnp_on else self._command.ModulePnp.OFF
-        # )
+        # if no module_id is specified, it will broadcast incoming pnp state
         if module_id is None:
             for curr_module_id in self._module_ids:
-                msg_to_write = self._command.set_module_state(
-                    curr_module_id, self._command.ModuleState.RUN, module_pnp_state
+                msg_to_write = self.__set_module_state(
+                    curr_module_id, Module.ModuleState.RUN, module_pnp_state
                 )
                 self._serial_write_q.put(msg_to_write)
+        # otherwise, it sets pnp state of the given module
         else:
-            msg_to_write = self._command.set_module_state(
-                module_id, self._command.ModuleState.RUN, module_pnp_state
+            msg_to_write = self.__set_module_state(
+                module_id, Module.ModuleState.RUN, module_pnp_state
             )
             self._serial_write_q.put(msg_to_write)
 
     def __append_hex(self, a, b):
+        # TODO: comment
         sizeof_b = 0
         while (b >> sizeof_b) > 0:
             sizeof_b += 1
         sizeof_b += sizeof_b % 4
         return (a << sizeof_b) | b
+
+    def __set_module_state(self, dst_id, module_state, pnp_state):
+        if type(module_state) is Module.ModuleState:
+            msg = dict()
+
+            msg["c"] = 0x09
+            msg["s"] = 0
+            msg["d"] = dst_id
+
+            state_bytes = bytearray(2)
+            state_bytes[0] = module_state.value
+            state_bytes[1] = pnp_state.value
+
+            msg["b"] = base64.b64encode(bytes(state_bytes)).decode("utf-8")
+            msg["l"] = 2
+
+            return json.dumps(msg, separators=(",", ":"))
+        else:
+            raise RuntimeError("The type of state is not ModuleState")
+
+    def init_modules(self):
+        BROADCAST_ID = 0xFFF
+
+        msg_to_send = self.__set_module_state(
+            BROADCAST_ID, Module.ModuleState.REBOOT, Module.ModulePnp.OFF
+        )
+        self._serial_write_q.put(msg_to_send)
+        self.__delay()
+
+        msg_to_send = self.__set_module_state(
+            BROADCAST_ID, Module.ModuleState.RUN, Module.ModulePnp.OFF
+        )
+        self._serial_write_q.put(msg_to_send)
+        self.__delay()
+
+        msg_to_send = self.__request_uuid(BROADCAST_ID)
+        self._serial_write_q.put(msg_to_send)
+        self.__delay()
+
+    def __delay(self):
+        time.sleep(1)
+
+    def __request_uuid(self, src_id, is_network_module=False):
+        BROADCAST_ID = 0xFFF
+
+        msg = dict()
+        msg["c"] = 0x28 if is_network_module else 0x08
+        msg["s"] = src_id
+        msg["d"] = BROADCAST_ID
+
+        id_bytes = bytearray(8)
+        id_bytes[0] = 0xFF
+        id_bytes[1] = 0x0F
+
+        msg["b"] = base64.b64encode(bytes(id_bytes)).decode("utf-8")
+        msg["l"] = 8
+
+        return json.dumps(msg, separators=(",", ":"))
