+++ modi/modi.py
@@ -1,15 +1,17 @@
 # -*- coding: utf-8 -*-
 
-"""Main module."""
-
-from __future__ import absolute_import
+"""Main MODI module."""
 
 import os
 import time
-import serial
+import json
+import base64
 
-from modi._processes import SerialProcess, ParserProcess, ExecutableThread
-from modi._command import Command
+from modi._serial_process import SerialProcess
+from modi._parser_process import ParserProcess
+from modi._executor_thread import ExecutorThread
+
+from modi.module.module import Module
 from modi.module import (
     button,
     dial,
@@ -25,43 +27,28 @@ from modi.module import (
     ultrasonic,
 )
 
-from multiprocessing import Process, Queue
+import multiprocessing
 
 
 class MODI:
     """
-    :param str port: MODI network module device name or ``None``.
-
-    :raises SerialException: In case the device can not be found or can not be configured.
-
-    The port is immediately opened on object creation, when a *port* is given. It is configured automatically when *port* is ``None`` and a successive call to :meth:`~modi.modi.MODI.open` is required.
-
-    *port* is a device name: depending on operating system. e.g. ``/dev/ttyUSB0`` on GNU/Linux or ``COM3`` on Windows.
-
     Example:
 
     >>> import modi
     >>> bundle = modi.MODI()
-
-    It can also be used with :meth:`modi.serial.list_ports`.
-
-    >>> import modi
-    >>> import modi.serial
-    >>> ports = modi.serial.list_ports() # [<serial.tools.list_ports_common.ListPortInfo object at 0x1026e95c0>]
-    >>> bundle = modi.MODI(ports[0].device)
     """
 
-    def __init__(self, port=None):
-        self._serial_read_q = Queue(100)
-        self._serial_write_q = Queue(100)
-        self._recv_q = Queue(100)
-        self._send_q = Queue(100)
+    def __init__(self):
+        self._serial_read_q = multiprocessing.Queue(100)
+        self._serial_write_q = multiprocessing.Queue(100)
 
-        self._src_ids = dict()
+        self._recv_q = multiprocessing.Queue(100)
+        self._send_q = multiprocessing.Queue(100)
+
         self._modules = list()
-        self._command = Command()
+        self._module_ids = dict()
 
-        self._ser_proc = SerialProcess(self._serial_read_q, self._serial_write_q, port)
+        self._ser_proc = SerialProcess(self._serial_read_q, self._serial_write_q)
         self._ser_proc.daemon = True
         self._ser_proc.start()
 
@@ -69,39 +56,14 @@ class MODI:
         self._par_proc.daemon = True
         self._par_proc.start()
 
-        self._exe_thrd = ExecutableThread(
-            self._serial_write_q,
-            self._recv_q,
-            self._src_ids,
-            self._modules,
-            self._command,
+        self._exe_thrd = ExecutorThread(
+            self._serial_write_q, self._recv_q, self._module_ids, self._modules
         )
         self._exe_thrd.daemon = True
         self._exe_thrd.start()
 
-        self.__init_modules()
-
-    def __init_modules(self):
-        BROADCAST_ID = 0xFFF
-
-        msg_to_send = self._command.set_module_state(
-            BROADCAST_ID, self._command.ModuleState.REBOOT, self._command.ModulePnp.OFF
-        )
-        self._serial_write_q.put(msg_to_send)
-        self.__delay()
-
-        msg_to_send = self._command.set_module_state(
-            BROADCAST_ID, self._command.ModuleState.RUN, self._command.ModulePnp.OFF
-        )
-        self._serial_write_q.put(msg_to_send)
-        self.__delay()
-
-        msg_to_send = self._command.request_uuid(BROADCAST_ID)
-        self._serial_write_q.put(msg_to_send)
-        self.__delay()
-
-    def __delay(self):
-        time.sleep(1)
+        # TODO: receive flag from executor thread
+        time.sleep(5)
 
     def exit(self):
         self._ser_proc.stop()
