+++ modi/module/__init__.py
@@ -2,19 +2,6 @@
 
 """Module package for pyMODI."""
 
-from __future__ import absolute_import
-
-from modi.module.button import Button
-from modi.module.dial import Dial
-from modi.module.display import Display
-from modi.module.env import Env
-from modi.module.gyro import Gyro
-from modi.module.ir import Ir
-from modi.module.mic import Mic
-from modi.module.motor import Motor
-from modi.module.speaker import Speaker
-from modi.module.ultrasonic import Ultrasonic
-
 __author__ = """Jinsung Ha"""
 __email__ = "jinsung@luxrobo.com"
 __version__ = "0.7.0"
