+++ modi/module/env.py
@@ -2,8 +2,6 @@
 
 """Env module."""
 
-from __future__ import absolute_import
-
 from enum import Enum
 from modi.module.module import InputModule
