+++ modi/module/motor.py
@@ -2,8 +2,6 @@
 
 """Motor module."""
 
-from __future__ import absolute_import
-
 from enum import Enum
 
 from modi.module.module import OutputModule
@@ -43,7 +41,7 @@ class Motor(OutputModule):
         """
         if value is not None:
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.INV.value,
                     (channel, mode, value, 0x00 if value >= 0 else 0xFFFF),
@@ -66,7 +64,7 @@ class Motor(OutputModule):
         """
         if degree is not None:
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.DEGREE.value,
                     (degree, self._get_property(self.PropertyType.FIRST_DEGREE), 0),
@@ -86,7 +84,7 @@ class Motor(OutputModule):
         """
         if degree is not None:
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.DEGREE,
                     (self.second_degree(), degree, 0),
@@ -106,7 +104,7 @@ class Motor(OutputModule):
         """
         if speed is not None:
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.SPEED.value,
                     (speed, self.first_speed(), 0),
@@ -126,7 +124,7 @@ class Motor(OutputModule):
         """
         if speed is not None:
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.SPEED.value,
                     (self.second_speed(), speed, 0),
@@ -146,7 +144,7 @@ class Motor(OutputModule):
         """
         if torque is not None:
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.TORQUE.value,
                     (torque, self.second_torque(), 0),
@@ -166,7 +164,7 @@ class Motor(OutputModule):
         """
         if torque is not None:
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.TORQUE.value,
                     (self.first_torque(), torque, 0),
@@ -193,7 +191,7 @@ class Motor(OutputModule):
                 self.second_torque() if second_torque is None else second_torque
             )
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.TORQUE.value,
                     (first_torque, second_torque, 0),
@@ -218,7 +216,7 @@ class Motor(OutputModule):
             first_speed = self.first_speed() if first_speed is None else first_speed
             second_speed = self.second_speed() if second_speed is None else second_speed
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.SPEED.value,
                     (first_speed, second_speed, 0),
@@ -245,7 +243,7 @@ class Motor(OutputModule):
                 self.second_degree() if second_degree is None else second_degree
             )
             self._serial_write_q.put(
-                self._command.set_property(
+                self._set_property(
                     self._module_id,
                     self.ControlType.DEGREE.value,
                     (first_degree, second_degree, 0),
