+++ modi/module/network.py
@@ -2,8 +2,6 @@
 
 """Network module."""
 
-from __future__ import absolute_import
-
 from enum import Enum
 
 from modi.module.module import SetupModule
